Moved files with like one thing in into AppUtils to tidy directory a bit
