+++ app/src/main/java/com/geecee/escapelauncher/utils/managers/FavouritesManager.kt
