+++ app/src/main/java/com/geecee/escapelauncher/utils/managers/HiddenAppsManager.kt
