+++ app/src/main/java/com/geecee/escapelauncher/utils/managers/OpenChallengesManager.kt
